# code simplification