Hi @bchevalier, I'm Paolo Lupo, this is my personal/hobby account :)

Talking about your PR https://github.com/liabru/matter-js/pull/533 I had some doubts about the need to define variables like: `positionAX = positionA.x`, `normalX = normal.x` etc.

Initially I thought this was *part* of the "trick", necessary to improve the performance some more, but I didn't like the idea (I'm a Java/C/C++ developer, I'm not *that* confident with Javascript).

Now... I don't know why you choose to follow that approach, if for a personal coding style or to follow liabru's example like `bodyA = collision.parentA` (I think he made this just for readability) or something else.

I removed some of those variables, and the performances didn't change (actually the get a little better but I think it's a random thing):

Stress example on chrome:

`original Resolver.solveVelocity (4800 samples): 1027 ms (18.5%)
new Resolver.solveVelocity (4800 samples): 640 ms (13.1%)
new Resolver.solveVelocity simplified (4800 samples): 662 ms (12.8%)
original Resolver.solveVelocity (4800 samples): 961 ms (19.0%)
new Resolver.solveVelocity (4800 samples): 622 ms (12.5%)
new Resolver.solveVelocity simplified (4800 samples): 578 ms (12.1%)
`

Complete logs:

[simplified.txt](https://github.com/bchevalier/matter-js/files/7516673/simplified.txt)

We saved 27 lines of code, but especially the code is more readable (less line of code used just to declare some aliases).

What do you think?