@bchevalier 2021-11-11
Hey @tylercorleone,
Well I think the changes you made could make sense if you assume that functions won't be de-optimized by object modification at run time.

So for instance in the following code:
```js
                var contact = pair.activeContacts[j],
                    offsetAX = contact.vertex.x - bodyA.position.x,
                    offsetAY = contact.vertex.y - bodyA.position.y,
                    offsetBX = contact.vertex.x - bodyB.position.x,
                    offsetBY = contact.vertex.y - bodyB.position.y,
```
Accessing the `x` property of the `vertex` property of `contact` could possibly made a single memory access with no condition on the types of contact or vertex. But due to the nature of JavaScript, if somewhere else in the code, the `contact` or `vertex` objects are added new properties dynamically. Then it would de-optimize this function even more in this simplified version.

So I guess what solution you will choose will depend on how sure you are that it won't happen.
Note that using typescript with very strict rules can almost ensure objects won't be de-optimized at run time. However I do not know any developer working with such strict rules that would avoid de-optimization completely and matterjs is not written in typescript.