@tylercorleone 2021-11-11
Hi @bchevalier,

by "if somewhere else in the code, the contact or vertex objects are added new properties dynamically" you mean inside the context of that function call (e.g. inside a nested function or something similar) or before the function itself being called?

I'm thinking about a test to replicate this scenario. Adding a random property to the vertex and/or contact each time the function is being called could be a good example? I could add a new field between two access to the same property and compare it with this version. I'm very focused in code readability and simplicity, but of course I would put it on a second plane if this would impact the performance.

I'm a little concerned for the fact liabru not responding.

I've seen a comparison of the performance of several physical engines and Matter Js was the worst one, but it is also the most used and the one with the best API, IMO, so I would like to improve it.